fix: make installed cli executable
