+++ packages/cli/package.json
@@ -12,7 +12,7 @@
   "scripts": {
     "typecheck": "tsc --noEmit -p tsconfig.json",
     "check:architecture": "node scripts/check-host-imports.mjs",
-    "bundle": "esbuild src/bin/texra.ts --bundle --platform=node --format=esm --target=node20 --external:fsevents --outfile=dist/bin/texra.js --banner:js='#!/usr/bin/env node'",
+    "bundle": "esbuild src/bin/texra.ts --bundle --platform=node --format=esm --target=node20 --external:fsevents --outfile=dist/bin/texra.js --banner:js='import { createRequire as __texraCreateRequire } from \"node:module\";const require = __texraCreateRequire(import.meta.url);'",
     "copy:resources": "node scripts/copy-resources.mjs",
     "build": "pnpm run typecheck && pnpm run check:architecture && pnpm run bundle && pnpm run copy:resources",
     "validate:run": "node scripts/validate-run.mjs"
