+++ packages/cli/src/runtime/logSinks.ts
@@ -1,3 +1,6 @@
+// Standard library imports
+import { createInterface } from 'node:readline/promises';
+
 // Local imports - logger
 import type { LogRecord, LogSink } from '@logger/structuredLogger';
 
@@ -57,6 +60,28 @@ export function writeTextStderr(text: string): void {
       stderrClosed = true;
     },
   );
+}
+
+export async function askCliQuestion(question: string): Promise<string> {
+  const prompt = createInterface({
+    input: process.stdin,
+    output: process.stderr,
+  });
+  try {
+    return await prompt.question(question);
+  } finally {
+    prompt.close();
+  }
+}
+
+export function createCliLineReader(
+  prompt: string,
+): ReturnType<typeof createInterface> {
+  return createInterface({
+    input: process.stdin,
+    output: process.stderr,
+    prompt,
+  });
 }
 
 export class StderrTextSink implements LogSink {
